Rolled out copyable_vector into a separate header.
