+++ inst/include/beachmat/utils/utils.h
@@ -16,22 +16,6 @@ namespace beachmat {
 template<class V>
 using const_col_indexed_info=std::tuple<size_t, Rcpp::IntegerVector::iterator, typename V::iterator>;
 
-// Definition of a copyable vector, to use in classes.
-
-template <class V> 
-struct copyable_holder {
-    copyable_holder(size_t n=0) : vec(n) {}
-    ~copyable_holder() = default;
-    copyable_holder(const copyable_holder& other) : vec(Rcpp::clone(other.vec)) {}
-    copyable_holder& operator=(const copyable_holder& other) { 
-        vec=Rcpp::clone(other.vec); 
-        return *this;
-    }
-    copyable_holder(copyable_holder&&) = default;
-    copyable_holder& operator=(copyable_holder&&) = default;
-    V vec;
-};
-
 /* String-related helper functions */
 
 inline std::string make_to_string(const Rcpp::RObject& str) {
